Move flow_content method to helper
This moves the flow_content method from the FlowRegistrationPresenter
to a seperate helper module ContentItemHelper. This is to reduce
complexity of the FlowRegistrationPresenter so it can be merged with the
FlowPresenter class.
